+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## Version 1.4.0 - 2024-03-12
+
+### Features
+
+- Adding support for reading and processing routes
+- Extracting co-reactant for ChemicalReaction class
+
+### Trivial changes
+
+- Making help for pipeline runner simpler
+
 ## Version 1.3.0 - 2024-01-09
 
 ### Features